refactor(ssh_dotnet): rename SSH_DotNet enum value + namespace/folder to SshDotNet
Unreleased feature, so the persisted format can change freely (no migration shim):
 - ProtocolType.SSH_DotNet -> ProtocolType.SshDotNet (the serialized Protocol="..."
   value changes accordingly; enum is name-serialized via GetAttributeAsEnum).
 - namespace + folder mRemoteNG[.|/]Connection.Protocol.SSH_DotNet -> SshDotNet
   (production + tests); updated every using/namespace.
 - SshDotNetDiagnostics log prefixes "[SSH_DotNet ...]" -> "[SshDotNet ...]"
   (and the matching TextLogMessageWriter test assertion).

Left untouched: Language.resx display strings (user-facing text, separate concern)
and the generated Language.Designer.cs. Other released protocols' enum values and
serialized strings are unchanged.
