+++ mRemoteNG/Connection/AbstractConnectionRecord.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.ComponentModel;
 using System.Drawing;
 using mRemoteNG.Connection.Protocol;
@@ -244,7 +244,7 @@ namespace mRemoteNG.Connection
         [LocalizedAttributes.LocalizedCategory(nameof(Language.Connection), 2),
          LocalizedAttributes.LocalizedDisplayName(nameof(Language.Username)),
          LocalizedAttributes.LocalizedDescription(nameof(Language.PropertyDescriptionUsername)),
-         AttributeUsedInProtocol(ProtocolType.RDP, ProtocolType.SSH1, ProtocolType.SSH2, ProtocolType.SSH_DotNet, ProtocolType.HTTP, ProtocolType.HTTPS, ProtocolType.IntApp)]
+         AttributeUsedInProtocol(ProtocolType.RDP, ProtocolType.SSH1, ProtocolType.SSH2, ProtocolType.SshDotNet, ProtocolType.HTTP, ProtocolType.HTTPS, ProtocolType.IntApp)]
         public virtual string Username
         {
             get => GetPropertyValue("Username", _username);
@@ -427,7 +427,7 @@ namespace mRemoteNG.Connection
         [LocalizedAttributes.LocalizedCategory(nameof(Language.Protocol), 3),
          LocalizedAttributes.LocalizedDisplayName(nameof(Language.SshPortForwardRules)),
          LocalizedAttributes.LocalizedDescription(nameof(Language.PropertyDescriptionSshPortForwardRules)),
-         AttributeUsedInProtocol(ProtocolType.SSH_DotNet)]
+         AttributeUsedInProtocol(ProtocolType.SshDotNet)]
         public string SSHDotNetPortForwardRules
         {
             get => GetPropertyValue("SSHDotNetPortForwardRules", _sshDotNetPortForwardRules);
