+++ mRemoteNG/Connection/Protocol/SshDotNet/SshAuthenticationProvider.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using Renci.SshNet;
 using Renci.SshNet.Common;
 
-namespace mRemoteNG.Connection.Protocol.SSH_DotNet
+namespace mRemoteNG.Connection.Protocol.SshDotNet
 {
     public static class SshAuthenticationProvider
     {
