+++ mRemoteNGTests/Connection/Protocol/SshDotNet/ProtocolSshDotNetTests.cs
@@ -1,9 +1,9 @@
-using mRemoteNG.Connection.Protocol.SSH_DotNet;
+using mRemoteNG.Connection.Protocol.SshDotNet;
 using NUnit.Framework;
 using System;
 using mRemoteNG.Connection;
 
-namespace mRemoteNGTests.Connection.Protocol.SSH_DotNet
+namespace mRemoteNGTests.Connection.Protocol.SshDotNet
 {
     [TestFixture]
     [Category("Unit")]
